FIAP/Cap 1 - Automação e inteligência na FarmTech Solutions/requirements.txt
streamlit
pandas
sqlalchemy
scikit-learn
joblib
pyyaml
pyserial